Compiler: fix ags4 fails script INT MIN
LONG_MIN gives a different value in non-MSVC x64 builds

+++ Compiler/script2/cc_symboltable.cpp
@@ -373,7 +373,7 @@ AGS::SymbolTable::SymbolTable()
     {
         Symbol const long_min_sym = Add("-2147483648");
         MakeEntryLiteral(long_min_sym);
-        entries[long_min_sym].LiteralD->Value = LONG_MIN;
+        entries[long_min_sym].LiteralD->Value = INT32_MIN;
         entries[long_min_sym].LiteralD->Vartype = kKW_Int;
     }
     {
